requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-dotenv
httpx
psycopg2-binary
pytest